+++ src/components/Header/MobileHeader/MenuBurger/MenuBurger.module.css
@@ -1,8 +1,9 @@
 .burgernav{
-    position: absolute;
     width: 250px;
-    height: 900px;
+    max-height: 1040px;
     display: flex;
+    position: absolute;
+    z-index: 1001;
     flex-direction: column;
     background-color: var(--cardbg);
 }
@@ -10,23 +11,34 @@
     user-select: none;
     cursor: pointer;
 }
+.CloseBtn{
+    fill: var(--main-text);
+}
 .burgernav__title{
+    margin: 24px auto 0px 12px;
     display: flex;
     flex-direction: row;
     align-items: center;
-    margin-left: 25px;
-    margin-top: 25px;
     font-weight: 700;
     user-select: none;
     cursor: pointer;
     color: var(--main-text);
 }
+.burgernav__titleLink{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 16px;
+    user-select: none;
+}
 .burgernav__title .title{
-    margin-left: 14px;
+    margin-left: auto;
+    margin-right: auto;
     font-size: 24px;
     display: inline-block;
     color: var(--main-text);
 }
+
 .burgernav__title .title__span {
     font-size: 24px;
     color: var(--main-green);
@@ -35,6 +47,7 @@
 .burgernav__account{
     margin-left: 45px;
     margin-top: 26px;
+    margin-right: auto;
     display: flex;
     flex-direction: row;
     align-items: center;
@@ -74,17 +87,15 @@
     flex-direction: row;
     text-align: center;
     align-items: center;
-    padding: 16px 24px;
+    padding: 16px 0px 16px 24px;
     color: var(--main-green);
     transition: .3s ease-in-out;
 }.activeLink::after{
     display: flex;
-    position: absolute;
     content: ' ';
     height: 40px;
     width: 5px;
-    top: 5;
-    right: 0px;
+    margin-left: auto;
     background-color: var(--main-green);
     animation: activeLinkAfterAnimation .3s ease-in-out;
 }
@@ -179,7 +190,9 @@
     justify-content: center;
 }
 .burgernav__downside{
-    margin-top: 16px;
+    display: flex;
+    padding-top: 16px;
+    margin: 0px 0px 24px 0px;
     color: var(--main-text);
 }
 .burgernav__downside .list {
