+++ src/components/Header/PcHeader/PcHeader.module.css
@@ -3,6 +3,7 @@
     height: 160px;
     border-bottom: 1px solid #EEEEEE;
     position: relative;
+    z-index: 1000;
 }
 .header__container{
 }
@@ -17,13 +18,13 @@
 .header__logo{
     display: flex;
     align-items: center;
-    column-gap: 42px;
+    justify-content: space-between;
 }
 .header__logo h1{
     color: var(--main-text);
     font-weight: 700;
-    font-size: 36px;
-    line-height: 54px;
+    font-size: 24px;
+    line-height: 36px;
 }
 .header__logo span{
     color: var(--main-green);
@@ -137,6 +138,11 @@
 
 }
 
+@media only screen and (max-width: 400px) {
+    .header{
+        display: none;
+    }
+}
 
 @-webkit-keyframes ping {
     0% {
